# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiErrorResponse {
  error: string;
}

export interface AdminLoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface AdminUser {
  id: string;
  name: string;
  email: string;
  role: string;
  /** @nullable */
  country?: string | null;
}

export interface AdminAuthResult {
  token: string;
  user: AdminUser;
}

export interface AdminStatusCount {
  status: string;
  count: number;
}

export interface AdminConsultation {
  id: string;
  /** @nullable */
  serialNumber?: string | null;
  clientName: string;
  lawyerName: string;
  subject: string;
  scheduledDate: string;
  scheduledTime: string;
  status: string;
  type: string;
  price: number;
  paymentStatus: string;
}

export interface AdminOverview {
  totalUsers: number;
  totalClients: number;
  totalLawyers: number;
  totalOffices: number;
  totalConsultations: number;
  suspendedOffices: number;
  grossRevenue: number;
  pendingCommission: number;
  collectedCommission: number;
  revenueQatar: number;
  revenueJordan: number;
  consultationsByStatus: AdminStatusCount[];
  recentConsultations: AdminConsultation[];
}

export interface AdminLawyer {
  id: string;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  officeName?: string | null;
  consultationsCount: number;
  createdAt: string;
}

export interface AdminClient {
  id: string;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  country?: string | null;
  consultationsCount: number;
  createdAt: string;
}

export interface AdminOffice {
  id: string;
  name: string;
  /** @nullable */
  ownerName?: string | null;
  /** @nullable */
  country?: string | null;
  isSuspended: boolean;
  /** @nullable */
  suspensionReason?: string | null;
  debtThreshold: number;
  pendingCommission: number;
}

export interface AdminDueRow {
  /** @nullable */
  officeId?: string | null;
  officeName: string;
  /** @nullable */
  lawyerId?: string | null;
  lawyerName: string;
  isSuspended: boolean;
  debtThreshold: number;
  totalPendingCommission: number;
  totalCollectedCommission: number;
  pendingCount: number;
  totalGross: number;
  exceedsThreshold: boolean;
}

export interface CollectDuesInput {
  officeId?: string;
  lawyerId?: string;
  notes?: string;
}

export interface CollectDuesResult {
  ok: boolean;
  collectedCount: number;
  totalCollected: number;
  message: string;
}

export interface KillSwitchResult {
  ok: boolean;
  processed: number;
  suspended: string[];
  reinstated: string[];
}