# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      description: Authenticate an admin with email + password, returns a bearer token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAuthResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /admin/me:
    get:
      operationId: getAdminProfile
      tags: [admin]
      summary: Current admin profile
      description: Returns the profile of the authenticated admin.
      responses:
        "200":
          description: Admin profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
...
components:
  schemas:
...
    ApiErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    AdminLoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 1
    AdminUser:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
        country:
          type: ["string", "null"]
    AdminAuthResult:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/AdminUser"
...