# lib/api-zod/src/generated/types/adminAuthResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminUser } from './adminUser';

export interface AdminAuthResult {
  token: string;
  user: AdminUser;
}